jamiaccount: token must be a uint64_t
Messages tokens in Jami are on 64 bits. So sendSipMessage must
take a uint64_t instead a int.

Gitlab: #326

+++ src/jamidht/jamiaccount.cpp
@@ -3560,7 +3560,7 @@ bool
 JamiAccount::sendSIPMessage(SipConnection& conn,
                             const std::string& to,
                             void* ctx,
-                            int token,
+                            uint64_t token,
                             const std::map<std::string, std::string>& data,
                             pjsip_endpt_send_callback cb)
 {
